Move ratelimit code to module
