+++ src/sentry/ratelimits/__init__.py
@@ -2,7 +2,14 @@ from django.conf import settings
 
 from sentry.utils.services import LazyServiceWrapper
 
-__all__ = ("for_organization_member_invite", "RateLimiter")
+__all__ = (
+    "for_organization_member_invite",
+    "above_rate_limit_check",
+    "can_be_ratelimited",
+    "get_rate_limit_key",
+    "get_rate_limit_value",
+    "RateLimiter",
+)
 
 from .base import RateLimiter  # NOQA
 
@@ -11,4 +18,10 @@ backend = LazyServiceWrapper(
 )
 backend.expose(locals())
 
-from .utils import for_organization_member_invite
+from .utils import (
+    above_rate_limit_check,
+    can_be_ratelimited,
+    for_organization_member_invite,
+    get_rate_limit_key,
+    get_rate_limit_value,
+)
